fix: exceptions for controllers

+++ src/main/java/com/emdev/wallet/service/IAccountService.java
@@ -1,14 +1,17 @@
 package com.emdev.wallet.service;
 
+import com.emdev.wallet.exceptions.RequestException;
 import com.emdev.wallet.model.Account;
 import com.emdev.wallet.repository.AccountRepository;
 import com.emdev.wallet.user.User;
 import com.emdev.wallet.user.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -27,7 +30,13 @@ public class IAccountService implements AccountService{
 
     @Override
     public Account getAccount(Integer accountId) {
-        return accountRepository.findByAccountId(accountId).orElse(null);
+        Optional<Account> account = accountRepository.findByAccountId(accountId);
+        if (!account.isPresent()) {
+            throw new RequestException("The requested account does not exist", "P-404",
+                    HttpStatus.NOT_FOUND);
+        }
+        return account.get();
+
     }
 
     @Override
@@ -41,7 +50,12 @@ public class IAccountService implements AccountService{
 
     @Override
     public void deleteAccount(Integer accountId) {
-    accountRepository.deleteById(accountId);
+        Optional<Account> account = accountRepository.findById(accountId);
+        if (!account.isPresent()) {
+            throw new RequestException("The requested account does not exist", "P-404",
+                    HttpStatus.NOT_FOUND);
+        } else {
+            accountRepository.deleteById(accountId);
+        }
     }
-
 }
